login page styling improved greatly
